Module arrow_schema::datatype

Enums

DataType

The set of datatypes that are supported by this implementation of Apache Arrow.

IntervalUnit

YEAR_MONTH, DAY_TIME, MONTH_DAY_NANO interval in SQL style.

TimeUnit

An absolute length of time in seconds, milliseconds, microseconds or nanoseconds.

UnionMode

Constants

DECIMAL128_MAX_PRECISION

The maximum precision for DataType::Decimal128 values

DECIMAL128_MAX_SCALE

The maximum scale for DataType::Decimal128 values

DECIMAL256_MAX_PRECISION

The maximum precision for DataType::Decimal256 values

DECIMAL256_MAX_SCALE

The maximum scale for DataType::Decimal256 values

DECIMAL_DEFAULT_SCALE

The default scale for DataType::Decimal128 and DataType::Decimal256 values