arrow_schema

Module extension

Extension types.

Constants

EXTENSION_TYPE_METADATA_KEY

The metadata key for a serialized representation of the ExtensionType necessary to reconstruct the custom type.

EXTENSION_TYPE_NAME_KEY

The metadata key for the string name identifying an ExtensionType.

Traits

ExtensionType

Extension types.