arrow_schema::extension

Module canonical

Canonical extension types.

The Arrow columnar format allows defining extension types so as to extend standard Arrow data types with custom semantics. Often these semantics will be specific to a system or application. However, it is beneficial to share the definitions of well-known extension types so as to improve interoperability between different systems integrating Arrow columnar data.

https://arrow.apache.org/docs/format/CanonicalExtensions.html#format-canonical-extensions

Re-exports

pub use bool8::Bool8;

pub use fixed_shape_tensor::FixedShapeTensor;

pub use fixed_shape_tensor::FixedShapeTensorMetadata;

pub use json::Json;

pub use json::JsonMetadata;

pub use opaque::Opaque;

pub use opaque::OpaqueMetadata;

pub use uuid::Uuid;

pub use variable_shape_tensor::VariableShapeTensor;

pub use variable_shape_tensor::VariableShapeTensorMetadata;

Modules

bool8 🔒

8-bit Boolean

fixed_shape_tensor 🔒

FixedShapeTensor

json 🔒

JSON

opaque 🔒

Opaque

uuid 🔒

UUID

variable_shape_tensor 🔒

VariableShapeTensor

Enums

CanonicalExtensionType

Canonical extension types.