arrow_schema::extension::canonical::json

Struct Json

pub struct Json(JsonMetadata);

The extension type for JSON.

Extension name: arrow.json.

The storage type of this extension is String or LargeString or StringView. Only UTF-8 encoded JSON as specified in rfc8259 is supported.

This type does not have any parameters.

Metadata is either an empty string or a JSON string with an empty object. In the future, additional fields may be added, but they are not required to interpret the array.

https://arrow.apache.org/docs/format/CanonicalExtensions.html#json

Tuple Fields
0: JsonMetadata