Struct arrow_schema::SchemaBuilder

pub struct SchemaBuilder {
    fields: Vec<FieldRef>,
}

A builder to facilitate building a Schema from iteratively from FieldRef

Fields
fields: Vec<FieldRef>
Implementations

impl SchemaBuilder

pub fn new() -> Self

Creates a new empty SchemaBuilder

pub fn with_capacity(capacity: usize) -> Self

Creates a new empty SchemaBuilder with space for capacity fields

pub fn push(&mut self, field: impl Into<FieldRef>)

Appends a FieldRef to this SchemaBuilder without checking for collision

pub fn try_merge(&mut self, field: &FieldRef) -> Result<(), ArrowError>

Appends a FieldRef to this SchemaBuilder checking for collision

If an existing field exists with the same name, calls Field::try_merge

pub fn finish(self) -> Schema

Consume this SchemaBuilder yielding the final Schema