Crate arrow_schema

Arrow logical types

Modules

datatype 🔒

datatype_display 🔒

datatype_parse 🔒

error 🔒

Defines ArrowError for representing failures in various Arrow operations.

extension

Extension types.

ffi

Contains declarations to bind to the C Data Interface.

field 🔒

fields 🔒

schema 🔒

Structs

Field

Describes a single column in a Schema.

Fields

A cheaply cloneable, owned slice of FieldRef

Schema

Describes the meta-data of an ordered sequence of relative types.

SchemaBuilder

A builder to facilitate building a Schema from iteratively from FieldRef

SortOptions

Options that define the sort order of a given column

UnionFields

A cheaply cloneable, owned collection of FieldRef and their corresponding type ids

Enums

ArrowError

Many different operations in the arrow crate return this error type.

DataType

Datatypes supported by this implementation of Apache Arrow.

IntervalUnit

YEAR_MONTH, DAY_TIME, MONTH_DAY_NANO interval in SQL style.

TimeUnit

An absolute length of time in seconds, milliseconds, microseconds or nanoseconds.

UnionMode

Sparse or Dense union layouts

Constants

DECIMAL32_DEFAULT_SCALE

The default scale for DataType::Decimal32 values

DECIMAL32_MAX_PRECISION

The maximum precision for DataType::Decimal32 values

DECIMAL32_MAX_SCALE

The maximum scale for DataType::Decimal32 values

DECIMAL64_DEFAULT_SCALE

The default scale for DataType::Decimal64 values

DECIMAL64_MAX_PRECISION

The maximum precision for DataType::Decimal64 values

DECIMAL64_MAX_SCALE

The maximum scale for DataType::Decimal64 values

DECIMAL128_MAX_PRECISION

The maximum precision for DataType::Decimal128 values

DECIMAL128_MAX_SCALE

The maximum scale for DataType::Decimal128 values

DECIMAL256_MAX_PRECISION

The maximum precision for DataType::Decimal256 values

DECIMAL256_MAX_SCALE

The maximum scale for DataType::Decimal256 values

DECIMAL_DEFAULT_SCALE

The default scale for DataType::Decimal128 and DataType::Decimal256 values

Type Aliases

FieldRef

A reference counted Field

SchemaRef

A reference-counted reference to a Schema.