arrow_schema

Module extension

Extension types.

This module is experimental. There might be breaking changes between minor releases.

Modules

canonical ๐

Canonical extension types.

Structs

Bool8

The extension type for 8-bit Boolean.

FixedShapeTensor

The extension type for fixed shape tensor.

FixedShapeTensorMetadata

Extension type metadata for FixedShapeTensor.

Json

The extension type for JSON.

JsonMetadata

Extension type metadata for Json.

Opaque

The extension type for Opaque.

OpaqueMetadata

Extension type metadata for Opaque.

Uuid

The extension type for UUID.

VariableShapeTensor

The extension type for VariableShapeTensor.

VariableShapeTensorMetadata

Extension type metadata for VariableShapeTensor.

Enums

CanonicalExtensionType

Canonical extension types.

Constants

EXTENSION_TYPE_METADATA_KEY

The metadata key for a serialized representation of the ExtensionType necessary to reconstruct the custom type.

EXTENSION_TYPE_NAME_KEY

The metadata key for the string name identifying an ExtensionType.

Traits

ExtensionType

Extension types.